Registration Backend
Created a user registration endpoint for the client to ping the backend whenever the registration form is submitted.

Takes the username, email, and password and calls the database to create the user. If the user is created then a status code of 201 meaning created is sent back. Otherwise, a status code of 500 is sent back.

Also cors was added to ensure cross platform api requests.

And urlencoded was used to ensure the body was parsable.