Register Frontend
Axios was added to make HTTP requests without the need for fetch or form native requests. Axios was imported. States set to have the event listen onChange, were automatically updated on user feedback. These states were used to submit a axios post request to the local server with the form information.

Added a Create Route in the App.jsx folder for later creations of posts.

Home page should be where the blogs go.

Layouts made 50% of the page useable for the content.

The register page did make one small change in that it calls the useNavigate() hook to navigate to the login page after a set timeout function when the response is authenticated.