Test Query
The query finds all users with the following test email. This email will be a query parameter in the database.js file. This is a test query to find a specific email in the database since all emails are unique, the account information can be found with validating the email.