Contact to Create and Deleted About
Reducing the scope of the project to have a create page and not a contact page or an about page. This is because it will crowd up the navigation menu and considering a forgot password section needs to implemented, static pages like this will get in the way of the more practical dynamic work.