Sessions Table
Created the sessions table for users to authenticate themselves with their session id. This table contains a unique id, a user id which is a foreign key, and a token variable character representing the session id.

Inserting into this table we have are dummy user ids and tokens.

Selecting all the sessions from the table.

Deleting all sessions form the table