Backend Cookies
Importing dotenv for the environment variables ( server info and client url ) are stored. Then dotenv.config() was called to ensure the dotenv file was being used.

cookieParser ensures cookies are parsed and sent correctly.

sessionID is created with bcrypt and the password then it is stored in a cookie and sent in the response.

Future: Still need to work on storing the sessionID in the database for future authentication references.