Delete Account
Users are now able to delete their account if they wish. Takes the user password as a input from the request and the session id from the cookie and checks if the password matches the session id from the database. If the passwords are a match then the session is deleted and the cookie is cleared. Then the user is deleted with the session id with a message that gets send if it was deleted.