Account Page FrontEnd
The account pages will hold user information such as the username, profile picture, bio, posts, and user settings such as logging out and deleting the account.

Functionalities for the logging out and deleting the account were implemented. The axios api was used to make calls to either log out or delete an account.