Storage + Avatar
new routes were created to get and post avatars to the server. The avatar is stored on the server under the public directory so they can be accessed via url. The server then stores the unique filename in the database so it can be accessed later under the users table.