Create Page
Form submission now reaches the server and all data is being received properly. Creating a database table that would hold the blog information is the next step in the process for creating the full blog UI for creating new blogs. This table would have to connect the user accounts to the blogs page such that a single person can have many blogs.