Creating Blogs
Saves blog information in database with the blogs table. The blogs table contains an id, and owner_id, a title, a picture, and the blog content. The create page makes axios requests using the multi-part form data object to send form data. The form data is targets the values and files listed in the create form through state variables.

The next thing to do is to redirect the user to the blog page that is created.