Multer File Upload
Now working with uploading blog images. Images are stored as filenames, with the blob stored in the public folder, so it is accessible across the internet. Personal information however, should not be stored in the public images folder. The next thing to do is to pull the data from the backend and create a visual interface for the user to read from. This will either be in a blog container then a blog page that will open from the preview container.