Blogs leading to account pages
attempting create blogs on the main page that lead into the account page. This was done through joining the blogs and users table to create a larger table to select the appropriate data.

Next thing to do is to fix the account pages so that the account that is shown is the correct account that created the blog.