Followers
Adding Followers to the Profile Page so users can follow, other users.

Created a followers table in the database to store the primary key of the following as well as two foreign keys: one being the user_id of the follower, and the other being the user id of the followed account.

This allows for selecting all the users who followed an account by choosing where the user id that equals the followed id. Implementing the unfollow and the get user followers was then as straight forward as mentioned above.

This will be used to create a followers feed, and a followers list.