Fixing HTTP Cookies with useContext API
The problem with production code was that the user cookies were being set to http only, meaning they could not be accessed via the client side.

This means I needed to find another way to set the isLoggedIn state to be true. The way I worked around this was by using the createContext api to pass isLogged in to key points in the application such as the log in page and log out page to set the is logged in to be true or not.