Update Blog + Update Format
Edit the blog with a forum which calls the api to put the new requested data in the blog with the correct blog id. The updated time of the blog will also be changed on submit. The old data will be deleted such as the old media in the files of the public directory in the server. And the page will be refreshed to reload the data into

Home page will be separate from the feed page. The page itself is now embedded down 70px and to the left 250px to fit the page into the remaining space left by the navigation bars. This allows for greater centering and responsiveness.